fetch all spec lists
