+++ .github/workflows/e2e-allow-list-cleanup.yml
@@ -1,4 +1,4 @@
-name: E2E Allow List Cleanup
+name: Allow List Cleanup
 
 on:
   # schedule:
@@ -6,8 +6,8 @@ on:
   push
 
 jobs:
-  fetch-e2e-allow-list:
-    name: Fetch E2E Allow List
+  fetch-allow-lists:
+    name: Fetch Allow Lists
     runs-on: ubuntu-latest
     outputs:
       allow_list: ${{ steps.allow-list-output.outputs.list }}
@@ -41,7 +41,7 @@ jobs:
           env_variable_name: VA_VSP_BOT_GITHUB_TOKEN
 
       - name: Get list of existing specs in vets-website
-        run: node script/github-actions/export-cypress-spec-list.js
+        run: node script/github-actions/export-test-lists.js
 
         # exports spec list and assigns to environmental variable at this step
       - name: Set Spec List Output
@@ -68,51 +68,64 @@ jobs:
           TEST_TYPE: unit_test
       # exports allow list and assigns to environmental variable at this step
 
-      - name: Archive e2e allow list
+      - name: Archive E2E allow list
         if: ${{ always() }}
         uses: actions/upload-artifact@v3
         with:
           name: e2e-allow-list
           path: qa-standards-dashboard-data/e2e_allow_list.json
 
-  clean-up-allow-list:
-    name: Clean up E2E Allow List
-    runs-on: ubuntu-latest
-    needs: [fetch-e2e-allow-list]
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Configure AWS credentials
-        uses: aws-actions/configure-aws-credentials@v1
+      - name: Archive Unit Test allow list
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v3
         with:
-          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
-          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          aws-region: us-gov-west-1
+          name: unit-test-allow-list
+          path: qa-standards-dashboard-data/unit_test_allow_list.json
 
-      - name: Get va-vsp-bot token
-        uses: marvinpinto/action-inject-ssm-secrets@v1.2.1
-        with:
-          ssm_parameter: /devops/VA_VSP_BOT_GITHUB_TOKEN
-          env_variable_name: VA_VSP_BOT_GITHUB_TOKEN
-
-      - name: Set Up BigQuery Creds
-        uses: ./.github/workflows/configure-bigquery
-
-      - name: Init Dashboard Data Repo
-        # uses: department-of-veterans-affairs/platform-release-tools-actions/init-data-repo@main
-        uses: ./.github/workflows/init-data-repo
-
-      - name: Download Allow List
-        uses: actions/download-artifact@v3
+      - name: Archive E2E spec list
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v3
         with:
           name: e2e-allow-list
-          path: ./qa-standards-dashboard-data
+          path: qa-standards-dashboard-data/e2e_spec_list.json
 
-      - name: Run cleanup script
-        run: yarn cleanup-allow-list
-        working-directory: qa-standards-dashboard-data
-        env:
-          ALL_SPECS: ${{ needs.fetch-e2e-allow-list.outputs.all_specs }}
-          TEST_TYPE: e2e
+      - name: Archive Unit Test spec list
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v3
+        with:
+          name: e2e-allow-list
+          path: qa-standards-dashboard-data/unit_test_spec_list.json
+  # clean-up-allow-list:
+  #   name: Clean up E2E Allow List
+  #   runs-on: ubuntu-latest
+  #   needs: [fetch-e2e-allow-list]
+  #   steps:
+  #     - name: Checkout
+  #       uses: actions/checkout@v3
+  #     - name: Configure AWS credentials
+  #       uses: aws-actions/configure-aws-credentials@v1
+  #       with:
+  #         aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
+  #         aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+  #         aws-region: us-gov-west-1
+  #     - name: Get va-vsp-bot token
+  #       uses: marvinpinto/action-inject-ssm-secrets@v1.2.1
+  #       with:
+  #         ssm_parameter: /devops/VA_VSP_BOT_GITHUB_TOKEN
+  #         env_variable_name: VA_VSP_BOT_GITHUB_TOKEN
+  #     - name: Set Up BigQuery Creds
+  #       uses: ./.github/workflows/configure-bigquery
+  #     - name: Init Dashboard Data Repo
+  #       # uses: department-of-veterans-affairs/platform-release-tools-actions/init-data-repo@main
+  #       uses: ./.github/workflows/init-data-repo
+  #     - name: Download Allow List
+  #       uses: actions/download-artifact@v3
+  #       with:
+  #         name: e2e-allow-list
+  #         path: ./qa-standards-dashboard-data
+  #     - name: Run cleanup script
+  #       run: yarn cleanup-allow-list
+  #       working-directory: qa-standards-dashboard-data
+  #       env:
+  #         ALL_SPECS: ${{ needs.fetch-e2e-allow-list.outputs.all_specs }}
+  #         TEST_TYPE: e2e
