--- script/github-actions/export-cypress-spec-list.js
@@ -1,18 +1,0 @@
-const glob = require('glob');
-const core = require('@actions/core');
-
-const countCySpecs = () => {
-  return new Promise((resolve, reject) => {
-    glob('**/*.cypress.spec.js?(x)', (error, files) => {
-      if (error) {
-        reject(error);
-        return;
-      }
-
-      resolve(files);
-    });
-  });
-};
-countCySpecs().then(files => {
-  core.exportVariable('ALL_CYPRESS_SPECS', files);
-});
